Geocoders' logger is set in the Railtie
ActiveRecord::Base.logger needs to be instanciated
before we set it to Geokit::Geocoders.logger

Maybe we could find how to override this in an environment variable
to be able to set our own logger for the geocoder.
For the moment it can only be changed in the app with the setter method.